Signup issues resolved and Product add edit complete. login issue for inactive users resolved

+++ angular-front/src/app/components/admin-related/product-roster/product-roster.component.ts
@@ -1,11 +1,18 @@
 import {AfterViewInit, Component, Inject, OnInit, ViewChild} from '@angular/core';
-import {ProductService} from "../../../services/product.service";
-import {Product} from "../../../model/product";
-import {MatTableDataSource} from "@angular/material/table";
-import {MatSort} from "@angular/material/sort";
-import {MatPaginator} from "@angular/material/paginator";
-import {Router} from "@angular/router";
-import {DOCUMENT} from "@angular/common";
+import {ProductService} from '../../../services/product.service';
+import {Product} from '../../../model/product';
+import {MatTableDataSource} from '@angular/material/table';
+import {MatSort} from '@angular/material/sort';
+import {MatPaginator} from '@angular/material/paginator';
+import {Router} from '@angular/router';
+import {DOCUMENT} from '@angular/common';
+import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
+import {DialogElementsExampleDialog} from '../../user-related/signup/signup.component';
+
+export interface DialogData {
+  msg: string;
+}
+
 
 @Component({
   selector: 'app-product-roster',
@@ -14,15 +21,16 @@ import {DOCUMENT} from "@angular/common";
 })
 export class ProductRosterComponent implements OnInit, AfterViewInit {
 
+  msg: string[];
   dataSource = new MatTableDataSource();
   displayedColumns: string[] = ['image', 'name', 'price', 'categoryType',
-    'status', 'stock','action'];
+    'status', 'stock', 'action'];
 
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
 
-  constructor(private productService: ProductService, private router: Router,
-              @Inject(DOCUMENT) private _document: Document) { }
+  constructor(private productService: ProductService, private router: Router, public dialog: MatDialog,
+              @Inject(DOCUMENT) private document: Document) { }
 
   ngOnInit(): void {
     this.fetchAllProducts();
@@ -33,10 +41,20 @@ export class ProductRosterComponent implements OnInit, AfterViewInit {
     this.dataSource.sort = this.sort;
   }
 
+  openDialog(msg1:string): void {
+    const dialogRef = this.dialog.open(DialogElementsExampleDialog, {
+     height: '200px',
+     width: '600px',
+     data: {msg: msg1}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      this.refreshPage();
+    });
+  }
+
   fetchAllProducts() {
-
     this.productService.getAllElevated().subscribe(
-
       data => {
         this.dataSource.data = data;
       }, error => {
@@ -45,26 +63,22 @@ export class ProductRosterComponent implements OnInit, AfterViewInit {
     );
 
   }
-
   onDelete(product: Product) {
-
-    this.productService.deleteProduct(product.id).subscribe(
-
-      data => {
-
-        console.log(data);
-
-      }
-
-    );
-
-    this.refreshPage();
-
+    if (confirm('Are you sure to delete this product ' + product.name)) {
+      this.productService.deleteProduct(product.id).subscribe(
+        data => {
+          console.log(data);
+          this.openDialog('Product Deleted Successfully');
+        }, error => {
+         this.openDialog(error.error.text);
+        }
+      );
+    }
   }
 
   refreshPage() {
     this.router.navigateByUrl('/admin-product-roster', {skipLocationChange: false}).then(() =>
-      this._document.defaultView.location.reload());
+      this.document.defaultView.location.reload());
   }
 
 }
